Fix issue with creation of company PSCs
The self link on a company PSC was incorrect. In the links object it would create another object called self which would contain an item called self.
In CompanyPscs.java changed the Field annotation for links to data.links. Renamed the field to links and any references.
The creation of the PSCs would always create a corporate PSC when checking the count of all PSCs in the repository. Changed this to only check the number of PSCs created for the test company the PSCs are being created for.
Updated unit tests

BI-11762